Create add_bbr.py
This version of the script does the following additional things:

It checks if the Outline manager is installed by looking for the `outline-manager` executable in the `/usr/local/bin` directory. If it is not found, it prints an error message and exits.
<!--more-->

It checks if the Outline server is running before attempting to stop it. If the server is not running, it prints an error message and returns without doing anything.
<!--more-->

It reads the Outline server config file (`/etc/outline/config.json`) as a Python dictionary, modifies the dictionary to enable BBR, and writes the modified dictionary back to the config file.
<!--more-->


> To use this script, you can pass the name of the Outline server as a command line argument

```python
python add_bbr.py myserver
```